% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sectxt_info}
\alias{sectxt_info}
\title{Retrieve a data frame of security.txt keys/values}
\usage{
sectxt_info(x)
}
\arguments{
\item{x}{a parsed \code{security.txt} created with \code{\link[=sec_parse]{sec_parse()}}}
}
\value{
data frame
}
\description{
Retrieve a data frame of security.txt keys/values
}
