% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sectxt_url}
\alias{sectxt_url}
\title{Determine security.txt URL for a given site/URL}
\usage{
sectxt_url(url)
}
\arguments{
\item{url}{URL}
}
\value{
character vector
}
\description{
Provide any URL for a resource and retrieve the URL for
the \code{security.txt} file. Strips off extraneous URL
components and appends \code{.well-known/security.txt}.
}
\examples{
sectxt_url("https://securitytxt.org/this/that/the/other.html")
}
