% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.r
\name{sectxt_validate}
\alias{sectxt_validate}
\title{Validate a \code{security.txt} Web Security Policies file}
\usage{
sectxt_validate(x)
}
\arguments{
\item{x}{an object created with \code{\link[=sectxt]{sectxt()}}}
}
\value{
logical; \code{TRUE} if all fields match current IETF standard, \code{FALSE} if not
}
\description{
Validate a \code{security.txt} Web Security Policies file
}
